requirements.txt
Django==5.2.4
pre_commit==4.2.0
Faker==37.4.0
djangorestframework ==3.16.0
django-extensions ==4.1
drf-spectacular == 0.28.0
pandas == 2.3.2
openpyxl == 3.1.5
parameterized == 0.9.0